move `slice` types etc into their own, top-level module
