+++ src/par_iter/mod.rs
@@ -53,10 +53,6 @@ mod skip;
 pub use self::skip::Skip;
 mod take;
 pub use self::take::Take;
-mod slice;
-pub use self::slice::{SliceIter, ChunksIter};
-mod slice_mut;
-pub use self::slice_mut::{SliceIterMut, ChunksMutIter};
 mod map;
 pub use self::map::{Map, MapOp, MapFn, MapCloned, MapInspect};
 mod weight;
