+++ src/string.rs
@@ -1,3 +1,9 @@
+//! This module contains extension methods for `String` that expose
+//! parallel iterators, such as `par_split_whitespace()`. You will
+//! rarely need to interact with it directly, since if you add `use
+//! rayon::prelude::*` to your file, that will include the helper
+//! traits defined in this module.
+
 use par_iter::*;
 use par_iter::internal::*;
 use std::cmp::min;
